+++ packages/backend/src/actions/create-analysis.ts
@@ -10,8 +10,9 @@ export const createAnalysis = async (
   action: Analysis
 ): Promise<{ requestId: string }> => {
   try {
-    const [data, resourceValue, language] = await createAnalysisRequestData(action);
-
+    const [data, resourceValue, language] = await createAnalysisRequestData(
+      action
+    );
     const response = await axios.post<{ request_id: string }>(
       config.MEGALINTER_ANALYSIS_URL,
       data
@@ -22,7 +23,7 @@ export const createAnalysis = async (
     reportStore.set({
       resourceType: action.inputType,
       resourceValue,
-      language
+      language,
     });
     subscribeToMegaLinter(requestId, reportStore);
     return { requestId };