+++ CHANGELOG.md
@@ -13,6 +13,8 @@ _Can be run using `npm run codetotal:beta`_
 - Front-end
   - Add language label
   - Add missing icon for Java
+  - Show language only if it's available
+  - Detect language for file analysis
 - Back-End
   - Bug fix: SBOM packages not showing up in report page. Async parsing of packages information in SBOM module